python_azure_sdk_project_01/requirements.txt
# Do not include azure-functions-worker in this file
# The Python Worker is managed by the Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues
azure-functions
python-dotenv
tzdata==2022.2
azure.identity
azure.keyvault.secrets
